cpt/requirements.txt
six>=1.10.0, <1.15.0
conan>=1.7.0, <1.28.0
tabulate>=0.8.0, <0.9.0
